ClothesBackend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
aiomysql==0.2.0
python-multipart==0.0.9
Pillow==10.4.0
google-genai==0.3.0
python-dotenv==1.0.1
cryptography
sqladmin